+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Fixed ImmutableMap crash upon consuming nectar.
 - Fixed pack formats for built-in resource packs.
 - Fixed syncing of Lockdown effects.
+- Fixed `bovinesandbuttercups:lock_effect` criterion trigger not triggering.